+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of lizard-fancylayers
 0.2 (unreleased)
 ----------------
 
-- Add support for percentiles.
+- Add support for percentiles, colors.
 
 
 0.1 (2012-11-15)